requirements.txt
flask==2.3.3
flask_login==0.6.2
werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0